requirements.txt
boto3
aws-cdk-lib==2.139.0
constructs>=10.0.0,<11.0.0
sagemaker==2.217.0